+++ packages/meedu/lib/src/providers/simple/simple_provider.dart
@@ -1,6 +1,6 @@
 part of '../base_provider.dart';
 
-class SimpleProvider<T> extends BaseProvider<T> {
+class SimpleProvider<T> extends BaseProvider<T> with ListeneableProvider<T> {
   SimpleProvider(
     super.creator, {
     super.autoDispose = true,