+++ packages/meedu/lib/src/providers/state/state_provider.dart
@@ -1,6 +1,7 @@
 part of '../base_provider.dart';
 
-class StateProvider<T extends StateNotifier<S>, S> extends BaseProvider<T> {
+class StateProvider<T extends StateNotifier<S>, S> extends BaseProvider<T>
+    with ListeneableProvider<T> {
   StateProvider(
     super.creator, {
     super.autoDispose = true,